# lib/api-zod/src/generated/types/aiAdviceInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiAdviceInputFinancialStats } from "./aiAdviceInputFinancialStats";
import type { AiAdviceInputMode } from "./aiAdviceInputMode";

export interface AiAdviceInput {
  financialStats: AiAdviceInputFinancialStats;
  question: string;
  mode: AiAdviceInputMode;
}