# lib/api-zod/src/generated/types/simulationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SimulationInput {
  monthlyIncome: number;
  monthlyExpenses: number;
  monthlySavings: number;
  annualReturn: number;
  currentSavings: number;
  years: number;
}