# lib/api-zod/src/generated/types/simulationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { YearlyDataPoint } from "./yearlyDataPoint";

export interface SimulationResult {
  finalNetWorth: number;
  optimizedFinalNetWorth: number;
  monthlyCashFlow: number;
  savingsRate: number;
  yearlyData: YearlyDataPoint[];
}